Refactor TypedNotification protocol
Public changes:

- Expanded documentation surrounding the TypedNotification protocol.

Public API breaking changes:

- The associated `Sender` type of a `TypedNotification` and its property has
been renamed to `Object` to match the naming of the Foundation Notification
type.

- The static name property of a `TypedNotification` is now a `Notification.Name`
instead of a `String`.

Internal API changes:

- The `TypedNotification` protocol is no longer responsible for constructing a
Foundation notification representation of itself. This is now the responsibility
of the `TypedNotificationCenter` itself.
- The typed notification object is no longer stored in the user info dictionary
of a foundation notification object under the "noti" key. It is now stored under
a more unique key that's prefixed with an underscore to make it clear it's an
implementation detail.