Rework NotificationObserver type
The class is now publicly available (so users can actually conform to
`TypedNotificationCenter` now). It now functions like a Disposable type. It
wraps a block that is run when a instance is deallocated. A
`TypedNotificationCenter` has to put its observer removal logic inside this
block.

`TypedNotificationCenter` types no longer need to provide a `removeObserver`
method since this is handled by the observer class.