Rename NotificationObserver to NotificationObservation
I went for a walk and I think this new time better reflects what the class
actually is. It's _not_ the observer, it's just a general disposable type that
performs some action when it's deallocated to manage the lifetime of the
observation.

Also, this new name lines up with the NSKeyValueObservation type in Foundation
that accomplishes a similar thing.