Change framework mach-o type to a static library
This creates an issue for the demo playground as playgrounds only support
dynamically linked libraries. To get around this, I've created a new shared
scheme (TypedNotification [Playground]) and build configuration (Playground)
that builds a framework with a dynamic library

This will present a small issue for Carthage users because two frameworks will
be built. However, since Carthage (xcodebuild) defaults to using the Release
configuration, the same framework will be produced for both builds.